code improvements and add crud for settings

+++ app/Http/Controllers/CustomerController.php
@@ -68,13 +68,10 @@ class CustomerController extends Controller
             $toValidate['first_name'] = 'required';
             $toValidate['last_name']  = 'required';
         }
-        if ($request->type == 'company') {
-            $toValidate['company_name'] = 'required';
-        }
+        if ($request->type == 'company') $toValidate['company_name'] = 'required';
+
         $validator = Validator::make($request->all(), $toValidate);
-        if ($validator->fails()) {
-            return $this->errorResponse($validator->errors(), Response::HTTP_UNPROCESSABLE_ENTITY);
-        }
+        if ($validator->fails()) return $this->errorResponse($validator->errors(), Response::HTTP_UNPROCESSABLE_ENTITY);
 
         try {
             /** Save here */
@@ -87,15 +84,13 @@ class CustomerController extends Controller
                 $customer->first_name = $request->first_name;
                 $customer->last_name  = $request->last_name;
             }
-            if ($request->type == 'company') {
-                $customer->company_name = $request->company_name;
-            }
+            if ($request->type == 'company') $customer->company_name = $request->company_name;
             $customer->save();
+
+            return $this->successResponse($this->transformer->transform($customer), Response::HTTP_CREATED);
         } catch(\Exception $e) {
             return $this->errorResponse(['Error' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
-
-        return $this->successResponse($this->transformer->transform($customer), Response::HTTP_CREATED);
     }
 
     /**
@@ -106,12 +101,13 @@ class CustomerController extends Controller
      */
     public function show(Request $request, $id)
     {
-        $customer = Customer::where('id', $id)->where('organization_id', $this->auth->organization_id)->first();
-        if (!is_null($customer)) {
-            return $this->successResponse($this->transformer->transform($customer), Response::HTTP_OK);
-        }
+        try {
+            $customer = Customer::where('id', $id)->where('organization_id', $this->auth->organization_id)->first();
 
-        return $this->errorResponse(['Status' => 'Not Found'], Response::HTTP_NOT_FOUND);
+            return $this->successResponse($this->transformer->transform($customer), Response::HTTP_OK);
+        } catch(\Exception $e) {
+            return $this->errorResponse(['Error' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
+        }
     }
 
     /**
@@ -143,13 +139,10 @@ class CustomerController extends Controller
             $toValidate['first_name'] = 'required';
             $toValidate['last_name']  = 'required';
         }
-        if ($request->type == 'company') {
-            $toValidate['company_name'] = 'required';
-        }
+        if ($request->type == 'company') $toValidate['company_name'] = 'required';
+
         $validator = Validator::make($request->all(), $toValidate);
-        if ($validator->fails()) {
-            return $this->errorResponse($validator->errors(), Response::HTTP_UNPROCESSABLE_ENTITY);
-        }
+        if ($validator->fails()) return $this->errorResponse($validator->errors(), Response::HTTP_UNPROCESSABLE_ENTITY);
 
         try {
             /** Update here */
@@ -161,15 +154,13 @@ class CustomerController extends Controller
                 $customer->first_name = $request->first_name;
                 $customer->last_name  = $request->last_name;
             }
-            if ($request->type == 'company') {
-                $customer->company_name = $request->company_name;
-            }
+            if ($request->type == 'company') $customer->company_name = $request->company_name;
             $customer->save();
+
+            return $this->successResponse($this->transformer->transform($customer), Response::HTTP_OK);
         } catch(\Exception $e) {
             return $this->errorResponse(['Error' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
-
-        return $this->successResponse($this->transformer->transform($customer), Response::HTTP_OK);
     }
 
     /**
@@ -180,11 +171,12 @@ class CustomerController extends Controller
      */
     public function destroy($id)
     {
-        $customer = Customer::where('id', $id)->where('organization_id', $this->auth->organization_id)->first();
-        if (!is_null($customer)) {
-            $customer->delete();
-            return $this->successResponse(['Status' => 'Ok'], Response::HTTP_OK);
+        try {
+            $customer = Customer::where('id', $id)->where('organization_id', $this->auth->organization_id)->delete();
+
+            return $this->successResponse(['Success' => $customer ? true : false], Response::HTTP_OK);
+        } catch(\Exception $e) {
+            return $this->errorResponse(['Error' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
-        return $this->errorResponse(['Status' => 'Not Found'], Response::HTTP_NOT_FOUND);
     }
 }
